repackage, move dependencies into their own package
